+++ numba_dpex/core/types/__init__.py
@@ -5,6 +5,7 @@
 from .array_type import Array
 from .dpctl_types import DpctlSyclEvent, DpctlSyclQueue
 from .dpnp_ndarray_type import DpnpNdArray
+from .kernel_api_types.range_types import NdRangeType, RangeType
 from .numba_types_short_names import (
     b1,
     bool_,
@@ -26,7 +27,6 @@ from .numba_types_short_names import (
     uint64,
     void,
 )
-from .range_types import NdRangeType, RangeType
 from .usm_ndarray_type import USMNdArray
 
 usm_ndarray = USMNdArray
